Assets functions

This function returns the date, band and URL fields for each assets of an STACItemCollection object. For the URL you can add the GDAL library drivers for the following schemes: HTTP/HTTPS files, S3 (AWS S3) and GS (Google Cloud Storage).

Usage

assets_url(items, append_gdalvsi = TRUE)

# S3 method for STACItemCollection
assets_url(items, append_gdalvsi = TRUE)

# S3 method for STACItem
assets_url(items, append_gdalvsi = TRUE)

assets_select(items, asset_names = NULL, filter_fn = NULL)

# S3 method for STACItemCollection
assets_select(items, asset_names = NULL, filter_fn = NULL)

# S3 method for STACItem
assets_select(items, asset_names = NULL, filter_fn = NULL)

Arguments

items

a STACItemCollection object representing the result of /stac/search, /collections/{collectionId}/items.

append_gdalvsi

a logical if true, gdal drivers are included in the URL of each asset. The following schemes are supported: HTTP/HTTPS files, S3 (AWS S3) and GS (Google Cloud Storage).

asset_names

a character with the assets names to be filtered. If NULL (default) all assets will be returned..

filter_fn

a function that will be used to filter the attributes listed in the properties.

Value

a list with the attributes of date, bands and paths.

Examples

if (FALSE) {
 # STACItemCollection object
 stac_item <- stac("https://brazildatacube.dpi.inpe.br/stac/") %>%
  stac_search(collections = "CB4_64_16D_STK-1", limit = 100,
         datetime = "2017-08-01/2018-03-01",
         bbox = c(-48.206,-14.195,-45.067,-12.272)) %>%
  get_request() %>% items_fetch(progress = FALSE)

 stac_item %>% assets_url()
}