Endpoint functions

The collections function implements the WFS3 /collections and /collections/{collectionId} endpoints.

Each endpoint retrieves specific STAC objects:

/collections: Returns a list of STAC Collections published in the STAC service

/collections/{collectionId}: Returns a single STAC Collection object

Usage

collections(q, collection_id = NULL, limit = NULL)

Arguments

q

a rstac_query object expressing a STAC query criteria.

collection_id

a character collection id to be retrieved.

limit

an integer defining the maximum number of results to return. If not informed, it defaults to the service implementation.

Value

A rstac_query object with the subclass collections for /collections/ endpoint, or a collection_id subclass for /collections/{collection_id} endpoint, containing all search field parameters to be provided to STAC API web service.

See also

Examples

if (FALSE) { # \dontrun{
 stac("https://brazildatacube.dpi.inpe.br/stac/") %>%
   collections() %>%
   get_request()

 stac("https://brazildatacube.dpi.inpe.br/stac/") %>%
   collections(collection_id = "CB4-16D-2") %>%
   get_request()
} # }