Collections functions

These functions provide support to work with doc_collectionsobjects.

collections_length(): shows how many items there are in the doc_items object.

collections_matched(): shows how many items matched the search criteria.

collections_fetch(): request all STAC Items through pagination.

collections_next(): fetches a new page from STAC service.

Usage

collections_next(collections, ...)

collections_matched(collections, matched_field)

collections_length(collections)

collections_fetch(collections, ..., progress = TRUE, matched_field = NULL)

Arguments

collections

a doc_collections object.

...

additional arguments. See details.

matched_field

a character vector with the path where is the number of collections returned.

progress

a logical indicating if a progress bar must be shown or not. Defaults to TRUE.

Value

collections_length(): an integer value.

collections_matched(): returns an integer value if the STAC web server does support this extension. Otherwise returns NULL.

collections_fetch(): a doc_items with all matched items.

collections_next(): fetches a new page from STAC service.

Details

Ellipsis argument (...) appears in different items functions and has distinct purposes:

collections_fetch() and collections_next(): ellipsis is used to pass additional httr options to GET method, such as add_headers or set_cookies.

Examples

if (FALSE) { # \dontrun{
# doc_items object
stac("https://cmr.earthdata.nasa.gov/stac/LPCLOUD") |>
  collections() |>
  get_request() |>
  collections_fetch()
} # }