Printing functions

The print function covers all objects in the rstac package:

stac(): returns a doc_catalog document from /stac (v0.8.0) or / (v0.9.0 or v1.0.0) endpoint.

stac_search(): returns a doc_items document from /stac/search (v0.8.0) or /search (v0.9.0 or v1.0.0) endpoint containing all Items that match the provided search predicates.

collections(): implements the /collections and /collections/{collectionId} endpoints. The former returns a doc_collections document that lists all collections published by the server, and the later returns a single doc_collection document that describes a unique collection.

items(): retrieves a doc_items document from /collections/{collectionId}/items endpoint and a doc_item document from /collections/{collectionId}/items/{itemId} endpoints.

The rstac package objects visualization is based on markdown, a lightweight markup language. You can paste the output into any markdown editor for a better visualization.

Call print() function to print the rstac's objects. You can determine how many items will be printed using n parameter.

Usage

# S3 method for class 'rstac_query'
print(x, ...)

# S3 method for class 'doc_catalog'
print(x, ...)

# S3 method for class 'doc_collections'
print(x, n = 10, ...)

# S3 method for class 'doc_collection'
print(x, ...)

# S3 method for class 'doc_items'
print(x, n = 10, ..., tail = FALSE)

# S3 method for class 'doc_item'
print(x, ...)

# S3 method for class 'doc_queryables'
print(x, n = 10, ...)

# S3 method for class 'doc_conformance'
print(x, n = 10, ...)

# S3 method for class 'doc_link'
print(x, ...)

# S3 method for class 'doc_links'
print(x, n = 10, ...)

Arguments

x

either a rstac_query object expressing a STAC query criteria or any rstac_doc.

...

other parameters passed in the functions.

n

number of entries to print. Each object has its own rule of truncation: the doc_collection objects will print 10 links by default. If the object has less than 20 collections, all collections will be shown. In doc_items, 10 features will be printed by default. To show all entries, use n = Inf.

tail

A logical value indicating if last features in doc_items object must be show.

See also

Examples

if (FALSE) { # \dontrun{
 # doc_items object
 stac_item_collection <-
   stac("https://brazildatacube.dpi.inpe.br/stac/") %>%
   stac_search(collections = "CB4-16D-2",
          bbox = c(-47.02148, -17.35063, -42.53906, -12.98314),
          limit = 15) %>%
   get_request()

 print(stac_item_collection, n = 10)

 # doc_collections object
 stac_collection <-
     stac("https://brazildatacube.dpi.inpe.br/stac/") %>%
     collections() %>%
     get_request()

 print(stac_collection, n = 5)

 # rstac_query object
 obj_rstac <- stac("https://brazildatacube.dpi.inpe.br/stac/")

 print(obj_rstac)
} # }