R client library for STAC (rstac)

R client library for STAC is a R package that interfaces STAC API endpoints. For more information about the STAC specification, please consult the specification page (<http://stacspec.org>). This package supports the version 0.8.0 of the STAC specification.

The rstac functions

The rstac package provides two categories of functions: API endpoints and assets access.

The rstac functions can be divided in two groups, one those functions implementing STAC API endpoints and, other, those functions for data access (e.g. assets) and metadata organization.

STAC API endpoints functions

stac: implements STAC /stac endpoint.

stac_collections: implements /collections and /collections/{collectionId} WFS3 endpoints.

stac_items: implements /collections/{collectionId}/items and /collections/{collectionId}/items/{itemId} WFS3 endpoints.

stac_search: implements STAC /stac/search endpoint.

Data access and organization functions

get_request: makes HTTP GET requests to STAC web service.

post_request: makes HTTP POST requests to STAC web service.

items_matched: returns how many items matched the search criteria.

items_length: informs how many items are stored locally.

items_fetch: fetches all matched items from service.

items_group: group items by some property field.

assets_download: download all assets in batch.

Data types

The package implements two S3 classes, stac_items and stac_collection. These classes are regular lists representing the corresponding JSON STAC objects.