Endpoint functions

The stac_items function implements WFS3 /collections/{collectionId}/items, and /collections/{collectionId}/items/{itemId} endpoints (v0.8.0).

Each endpoint retrieves specific STAC objects:

/collections/{collectionId}/items: Returns a STAC Items collection (GeoJSON)

/collections/{collectionId}/items/{itemId}: Returns a STAC Item (GeoJSON Feature)

The endpoint /collections/{collectionId}/items accepts the same filters parameters of stac_search function.

stac_items(url, collection_id, item_id, datetime, bbox, limit, ...)

Arguments

url
A character informing the base url of a STAC web service.

collection_id
A character with a collection id to retrieve collection details.

item_id
A character with item id to be fetched. Only works if the collection_id is informed. This is equivalent to the endpoint /collections/{collectionId}/items/{itemId}.

datetime
Either a date-time or an interval. Date and time strings needs to conform RFC 3339. Intervals are expressed by separating two date-time strings by '/' character. Open intervals are expressed by using '..' in place of date-time.

Examples:

A date-time: "2018-02-12T23:20:50Z"

A closed interval: "2018-02-12T00:00:00Z/2018-03-18T12:31:12Z"

Open intervals: "2018-02-12T00:00:00Z/.." or "../2018-03-18T12:31:12Z"

Only features that have a datetime property that intersects the interval or date-time informed in datetime are selected.

bbox
Only features that have a geometry that intersects the bounding box are selected. The bounding box is provided as four or six numbers, depending on whether the coordinate reference system includes a vertical axis (elevation or depth):

Lower left corner, coordinate axis 1

Lower left corner, coordinate axis 2

Lower left corner, coordinate axis 3 (optional)

Upper right corner, coordinate axis 1

Upper right corner, coordinate axis 2

Upper right corner, coordinate axis 3 (optional)

The coordinate reference system of the values is WGS84 longitude/latitude (http://www.opengis.net/def/crs/OGC/1.3/CRS84). The values are in most cases the sequence of minimum longitude, minimum latitude, maximum longitude and maximum latitude. However, in cases where the box spans the antimeridian the first value (west-most box edge) is larger than the third value (east-most box edge).

limit
An integer defining the maximum number of results to return. If NULL it defaults to the service implementation. Defaults to 10.

...
Filter parameters. Accept the same filter parameters of stac_search function.

Value

A stac object containing all request parameters to be provided to stac_request.

Note

Param intersects is a character value expressing GeoJSON geometries objects as specified in RFC 7946. This param is not supported in current version.

See also

Examples

if (FALSE) {

stac_items("http://brazildatacube.dpi.inpe.br/bdc-stac/0.8.0",
           collection_id = "CB_64_16D_STK",
           bbox = c(-55.16335, -4.26325, -49.31739, -1.18355)) %>%
    get_request()

stac_items("http://brazildatacube.dpi.inpe.br/bdc-stac/0.8.0",
           collection_id = "MOD13Q1",
           item_id = "MOD13Q1.A2019241.h13v09.006.2019262164754") %>%
    get_request()
}