R client library for STAC (rstac)

Provides functions to access, search and download spacetime earth observation data via SpatioTemporal Asset Catalog (STAC). This package supports the version 1.0.0 of the STAC specification (https://github.com/radiantearth/stac-spec). For further details see Simoes et al. (2021) doi:10.1109/IGARSS47720.2021.9553518 .

The rstac functions

The rstac package provides two categories of functions: API endpoints and data access and organization.

STAC API endpoints functions

stac(): implements STAC /stac endpoint for version 0.8.1 or below, and / for versions 0.9.0 or higher.

collections(): implements /collections and /collections/{collectionId} WFS3 endpoints.

items(): implements /collections/{collectionId}/items and /collections/{collectionId}/items/{featureId} WFS3 endpoints.

stac_search(): implements STAC /stac/search endpoint for version 0.8.1 or below, and /search endpoint for versions 0.9.0 or higher.

Data access and organization functions

get_request(): makes HTTP GET requests to STAC web service.

post_request(): makes HTTP POST requests to STAC web service.

items_matched(): returns how many items matched the search criteria.

items_length(): informs how many items are stored locally.

items_fetch(): fetches all matched items from service.

assets_download(): download all assets in batch.

Data types

The package implements the follow S3 classes: STACItemCollection, STACItem, STACCatalog, STACCollectionList and STACCollection. These classes are regular lists representing the corresponding JSON STAC objects.

See also

Author

Maintainer: Brazil Data Cube Team brazildatacube@inpe.br

Other contributors:

National Institute for Space Research (INPE) [copyright holder]